requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
Werkzeug==3.0.1